pre-review commit
implements player inventory in order to count how many of each item the player has and fits Item script to use and modify these values as new items are picked up. also sets all items to scale non-linearly (this can be easily changed in the editor as wanted).
code is now ready for review and maybe implementation through code  review.